requirements.txt
Flask==0.12.2
Jinja2==2.9.6
MarkupSafe==1.0
Pillow==4.1.1
PyJWT==1.5.0
PySocks==1.6.7
Werkzeug==0.12.2
certifi==2017.4.17
chardet==3.0.3
click==6.7
idna==2.5
imageio==2.2.0
itsdangerous==0.24
numpy==1.12.1
olefile==0.44
pydub==0.19.0
python-dotenv==0.6.4
pytz==2017.2
qrcode==5.3
requests==2.17.3
six==1.10.0
twilio==6.3.0
urllib3==1.21.1
wsgiref==0.1.2
